Add 'with tempfile.TemporaryDirectory()' context manager using a template

+++ src/pylib.nim
@@ -144,13 +144,13 @@ template with_TemporaryDirectory*(statements: untyped): untyped =
   block:  # Error: redefinition.
     let name {.inject.} = getTempDir() / $rand(100_000..999_999)
     when not defined(release): echo name
-    discard existsOrCreateDir(name)
-    # Python people always try to use file.read() that wont exist on Nim.
-    template read(fileType: File): string {.used.} = fileType.readAll()
-    try: # defer: wont like top level,because is a template itself.
-      statements
-    finally:
-      try: removeDir(name) except: discard
+    if not existsOrCreateDir(name): # Returns true if directory already exists.
+      # Python people always try to use file.read() that wont exist on Nim.
+      template read(fileType: File): string {.used.} = fileType.readAll()
+      try: # defer: wont like top level,because is a template itself.
+        statements
+      finally:
+        try: removeDir(name) except: discard
 
 template pass*(_: any) = discard # pass 42
 template pass*() = discard       # pass()
